feat: entropy/adaptive lossless proof + zstd mount archive; fix make profile tests
- STUDYCASE_Q8_COMPRESSION_WALL: entropy wall ~1.0x on Q8_0 (3-direction proof)
- huffman_test: canonical Huffman order-0, lossless 100% (real SmolLM2)
- beam_compress_fixed: adaptive codec, stride-34 measure, 0.9896x real
- fgls_tensor_archive: zstd blob layer (KIS removed, was 1.7x expand)
- Makefile: auto-detect msys64 gcc (old C:/mingw64 8.1 broken)
- fgls_profile: KIS heuristic unique<=200 (align is_kis_candidate)
- geo_frame_seek: restore verify fn; tensor_track: uint16 pointer fix

+++ core/tensor_track.h
@@ -473,7 +473,7 @@ static inline TTStoreResult tt_store(EncCtx *enc_ctx,
     sr.frame_hi = fr.frame_hi;
 
     /* 2. Field position from enc */
-    uint32_t ring, wedge;
+    uint16_t ring, wedge;
     ft_enc_to_field(rec->enc, &ring, &wedge);
 
     /* 3. Pack into enclosure chunk (hexagon spread) */
